+++ domain/src/main/java/basecode/com/domain/model/network/response/InfoHomeResponse.kt
@@ -5,7 +5,7 @@ class InfoHomeResponse(val lstNewNews: MutableList<NewNewsModel> = mutableListOf
                        val lstNewEBook: MutableList<NewEBookModel> = mutableListOf(),
                        val lstCollectionRecommend: MutableList<CollectionRecommendModel> = mutableListOf())
 
-class NewNewsModel(val id: Int, val picture: String, val title: String, val content: String)
-class NewBooksModel(val id: Int, val title: String, val coverPicture: String, val author: String)
-class NewEBookModel(val id: Int, val title: String, val coverPicture: String, val author: String)
+class NewNewsModel(val id: Long, val picture: String, val title: String, val content: String)
+class NewBooksModel(val id: Long, val title: String, val coverPicture: String, val author: String)
+class NewEBookModel(val id: Long, val title: String, val coverPicture: String, val author: String)
 class CollectionRecommendModel(val id: Int, val name: String, val coverPicture: String)
