+++ ui/src/main/java/basecode/com/ui/features/home/viewholder/NewBookItemViewHolderModel.kt
@@ -2,4 +2,4 @@ package basecode.com.ui.features.home.viewholder
 
 import basecode.com.ui.base.listview.model.ViewHolderModel
 
-class NewBookItemViewHolderModel(val id: Int, val title: String, val coverPicture: String, val author: String): ViewHolderModel
+class NewBookItemViewHolderModel(val id: Long, val title: String, val coverPicture: String, val author: String): ViewHolderModel
