+++ ui/src/main/java/basecode/com/ui/features/searchbook/BookViewHolderModel.kt
@@ -2,4 +2,4 @@ package basecode.com.ui.features.searchbook
 
 import basecode.com.ui.base.listview.model.ViewHolderModel
 
-class BookViewHolderModel(val id: Int, val name: String, val photo: String, val author: String, var publishedYear: String): ViewHolderModel
+class BookViewHolderModel(val id: Long, val name: String, val photo: String, val author: String, var publishedYear: String): ViewHolderModel
